lots of small changes. USB UART now works

+++ Node-Beaver/Node-Beaver.cydsn/data.h
@@ -11,7 +11,7 @@ typedef struct
 {
   uint32_t time;
   uint16_t type;
-  uint32_t value;
+  uint64_t value;
 } DataPacket;
 
 #endif
